pyproject.toml
[project]
name = "nonebot-plugin-osumania-toolkit"
version = "1.1.0"
description = "a nonebot plugin with some useful osu!mania utilities"
readme = "README.md"
requires-python = ">=3.10,<3.13"
license = {text = "MIT"}
authors = [
    {name = "LeoBlackMT", email = "Leo_BlackMT@qq.com"}
]

dependencies = [
    "nonebot2>=2.3.1",
    "nonebot-adapter-onebot>=2.0.0",
    "nonebot-plugin-localstore>=0.3.0",
    "nonebot-plugin-htmlkit>=0.1.0rc5,<0.2.0",
    "aiohttp>=3.8.0",
    "Pillow>=10.0.0",
    "numpy>=1.26.0",
    "onnxruntime>=1.23.2,<1.24.0",
    "scipy>=1.7.0",
    "statsmodels>=0.14.0",
    "matplotlib>=3.5.0",
    "pandas>=2.2.0,<3.0.0",
]

[project.urls]
Homepage = "https://github.com/LeoBlackMT/nonebot-plugin-osumania-toolkit"

[project.entry-points."nonebot.plugin"]
nonebot-plugin-osumania-toolkit = "nonebot_plugin_osumania_toolkit"
nonebot_plugin_osumania_toolkit = "nonebot_plugin_osumania_toolkit"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.build]
includes = ["src"]

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.4.2",
    "isort>=5.13.2",
    "ruff>=0.4.6",
]

[tool.pdm.scripts]
format = { composite = ["isort . ","black . ","ruff check ."] }
test = "python -c 'print(\">>> Just \\\"pytest -W ignore -s\\\" when you complete your testsuite\")'"


[tool.black]
line-length = 90
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
extend-exclude = '''
'''

[tool.isort]
profile = "black"
line_length = 90
length_sort = true
skip_gitignore = true
force_sort_within_sections = true
extra_standard_library = ["typing_extensions"]

[tool.ruff]
line-length = 90
target-version = "py310"

[tool.ruff.lint]
select = ["E", "W", "F", "UP", "C", "T", "PYI", "PT", "Q"]
ignore = ["C901", "T201", "E731", "E402"]

[tool.pyright]
pythonVersion = "3.10"
pythonPlatform = "All"
typeCheckingMode = "basic"

[tool.nonebot]
adapters = [

]
plugin_dirs = ["src"]
builtin_plugins = []